final-touches - add product image upload - add more validations - add error handling for purchased product - apply more UI (#24)

+++ src/main/java/com/waa/dormart/controllers/seller/SellerProductController.java
@@ -3,17 +3,25 @@ package com.waa.dormart.controllers.seller;
 import com.waa.dormart.exceptions.HttpException;
 import com.waa.dormart.exceptions.UnuthorizedException;
 import com.waa.dormart.models.Product;
+import com.waa.dormart.models.ShoppingOrder;
 import com.waa.dormart.models.User;
 import com.waa.dormart.services.CategoryService;
 import com.waa.dormart.services.ProductService;
+import com.waa.dormart.services.ShoppingOrderService;
+import com.waa.dormart.services.StorageService;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.ObjectError;
+import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.Valid;
+import java.time.LocalDate;
+import java.time.ZonedDateTime;
 import java.util.List;
 
 @Controller
@@ -21,10 +29,17 @@ import java.util.List;
 public class SellerProductController {
     private ProductService productService;
     private CategoryService categoryService;
+    private StorageService storageService;
+    private ShoppingOrderService shoppingOrderService;
 
-    public SellerProductController(ProductService productService, CategoryService categoryService) {
+    public SellerProductController(ProductService productService,
+                                   CategoryService categoryService,
+                                   StorageService storageService,
+                                   ShoppingOrderService shoppingOrderService) {
         this.productService = productService;
         this.categoryService = categoryService;
+        this.storageService = storageService;
+        this.shoppingOrderService = shoppingOrderService;
     }
 
     @GetMapping
@@ -36,6 +51,10 @@ public class SellerProductController {
 
     @PostMapping("{id}/delete")
     public String deleteProduct(@PathVariable("id") Long productId) {
+        List<ShoppingOrder> orders = shoppingOrderService.getDeliveredProductOrders(productId);
+        if (orders.size() > 0) {
+            throw new UnuthorizedException(HttpStatus.UNAUTHORIZED, "Product cannot be deleted because it is already purchased");
+        }
         productService.deleteProduct(productId);
         return "redirect:/seller/products";
     }
@@ -47,26 +66,37 @@ public class SellerProductController {
     }
 
     @GetMapping("new")
-    public String getProductForm(@AuthenticationPrincipal User seller, @ModelAttribute Product product, Model model) {
+    public String getProductForm(@AuthenticationPrincipal User seller,
+                                 @ModelAttribute Product product, Model model) {
         product.setSeller(seller);
         model.addAttribute(product);
         model.addAttribute("categories", categoryService.findAll());
         return "sellers/products/new";
     }
 
-    @PostMapping
+    @PostMapping("new")
     public String save(@AuthenticationPrincipal User seller,
                        @Valid @ModelAttribute Product product,
                        BindingResult result,
-                       Model model) {
+                       Model model,
+                       @RequestPart("image") MultipartFile image) {
         if (!seller.getActive()) {
             throw (HttpException) new UnuthorizedException(HttpStatus.UNAUTHORIZED, "You cannot post any product until the admin approves your account");
         }
         product.setSeller(seller);
+
+        if (image == null) {
+            result.addError(new ObjectError("image", "{model.notBlank.error}"));
+        }
+
         if (result.hasErrors()) {
             model.addAttribute("categories", categoryService.findAll());
             return "sellers/products/new";
         }
+        String filename = String.format("%s.%s", System.currentTimeMillis(), image.getContentType().split("/")[1]);
+        storageService.store(image, filename);
+
+        product.setImage(filename);
 
         this.productService.saveProduct(product);
 
@@ -81,10 +111,11 @@ public class SellerProductController {
     }
 
     @PostMapping("{id}/edit")
-    public String udpateProduct(@AuthenticationPrincipal User seller,
+    public String updateProduct(@AuthenticationPrincipal User seller,
                        @Valid @ModelAttribute Product product,
                        BindingResult result,
                        Model model,
+                       @RequestPart("image") MultipartFile image,
                        @PathVariable Long id) {
         product.setId(id);
         product.setSeller(seller);
@@ -93,8 +124,20 @@ public class SellerProductController {
             return "sellers/products/new";
         }
 
+        if (image == null) {
+            String filename = String.format("%s.%s", System.currentTimeMillis(), image.getContentType().split("/")[1]);
+            storageService.store(image, filename);
+
+            product.setImage(filename);
+        }
+
         this.productService.saveProduct(product);
 
         return "redirect:/seller/products";
+    }
+
+    @InitBinder
+    public void initializeBinder(WebDataBinder binder) {
+        binder.setDisallowedFields("image");
     }
 }
